Move BayesianRecommender to new subpackage
