+++ baybe/recommenders/bayesian/__init__.py
@@ -1,0 +1,1 @@
+"""Bayesian recommenders."""
