+++ tests/serialization/test_naive_hybrid_serialization.py
@@ -4,10 +4,10 @@ import pytest
 
 from baybe.campaign import Campaign
 from baybe.recommenders.bayesian import (
-    BayesianRecommender,
     NaiveHybridRecommender,
     SequentialGreedyRecommender,
 )
+from baybe.recommenders.bayesian.base import BayesianRecommender
 from baybe.recommenders.nonpredictive.base import NonPredictiveRecommender
 from baybe.searchspace import SearchSpaceType
 from baybe.utils.basic import get_subclasses
